Use latest version of mocha (v1.11.0)
Since v1.10.0, when an invocation matches an expectation which was
instructed to yield using Expectation#yields or
Expectation#multiple_yields, the stubbed method will *always* yield
whether or not the invocation supplies a block.

The quality_xxx method on all of the Tool subclasses (except Shellcheck)
invokes Runner#ratchet_quality_cmd with a block, but the same method on
Shellcheck invokes it with no block. So we need to adjust the
expectation accordingly.

I don't know enough about this project to know whether there's a better
way to identify the Shellcheck Tool subclass, but this solves the
immediate problem when upgrading mocha.